+++ src/components/MultiSeriesBarChart/components/BarGroup/BarGroup.tsx
@@ -32,6 +32,7 @@ interface Props {
   onFocus: (index: number) => void;
   hasRoundedCorners: boolean;
   isAnimated?: boolean;
+  allValuesNegative?: boolean;
 }
 
 export function BarGroup({
@@ -47,6 +48,7 @@ export function BarGroup({
   hasRoundedCorners,
   isSubdued,
   isAnimated = false,
+  allValuesNegative = false,
 }: Props) {
   const {prefersReducedMotion} = usePrefersReducedMotion();
   const barWidth = width / data.length - BAR_SPACING;
@@ -114,6 +116,7 @@ export function BarGroup({
                 role={ariaEnabledBar ? 'img' : undefined}
                 ariaLabel={ariaEnabledBar ? ariaLabel : undefined}
                 hasRoundedCorners={hasRoundedCorners}
+                allValuesNegative={allValuesNegative}
               />
             </g>
           );